Segment based on histogram peaks and random walker
After seeing that the histograms of WT and KO are different we thought
that there's some hidden meaning in that :)
Based on peaks detected on the gray value histogram we initialize a
random walker segmentation and plot the segmented volume of the fishes.
This segmented volume seems to correspond to their muscle volume quite
well.
This is the status of the notebook for further discussion with Carolina
and Nadia.